[!] Bug fix: Updated cookie handling
Cookies are now stored in an array in the "set-cookie" index of the
"@meta" associative array (hash)

This commit patches an error introduced in commit:

	https://github.com/bcoles/WhatWeb/commit/
	f525a5551264c6b69259ebaed8cc710d65a0be99

The aforementioned commit was originally intended to remove @cookies
however cookie detection failed when more than one cookie was present.

Cookies are now handled properly without using the @cookies array.